# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.ResetPasswordCreatable:
    properties:
      email:
        type: string
    type: object
...
  models.SignIn:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  models.SignUp:
    properties:
      email:
        type: string
      fullName:
        type: string
      gender:
        type: boolean
      password:
        type: string
      roleId:
        type: integer
      tel:
        type: string
    type: object
...
  models.response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      statusCode:
        type: integer
    type: object
host: localhost:3001
info:
  contact:
    email: dso.intern.xuanhoa@gmail.com
    name: Xuan Hoa Le
  description: Go Food Delivery CMS - Swagger API Discovery
  title: Go Food Delivery CMS - Swagger API Discovery
  version: "1.0"
paths:
...
  /auth/me:
    get:
      consumes:
      - application/json
      description: Me
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.response'
      summary: Me
      tags:
      - auth
  /auth/reset-password:
    patch:
      consumes:
      - application/json
      description: Reset Password
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Reset Password
        in: body
        name: resetPassword
        required: true
        schema:
          $ref: '#/definitions/models.ResetPasswordCreatable'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.response'
      summary: Reset Password
      tags:
      - auth
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: SignIn
      parameters:
      - description: Sign In
        in: body
        name: signIn
        required: true
        schema:
          $ref: '#/definitions/models.SignIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.response'
      summary: SignIn
      tags:
      - auth
  /auth/sign-up:
    post:
      consumes:
      - application/json
      description: Sign up
      parameters:
      - description: Sign Up
        in: body
        name: signUp
        required: true
        schema:
          $ref: '#/definitions/models.SignUp'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.response'
      summary: Sign up
      tags:
      - auth
  /bills/compensated/:
    patch:
      consumes:
      - application/json
      description: Compensated Bill
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Order Id
        in: query
        name: orderId
        required: true
        type: integer
      - description: Bill Id
        in: query
        name: billId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.response'
      summary: Compensated Bill
      tags:
      - bills
  /bills/finished/{billId}:
    patch:
      consumes:
      - application/json
      description: Finish Bill
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Bill Id
        in: path
        name: billId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.response'
      summary: Finish Bill
      tags:
      - bills
...
swagger: "2.0"